# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ihyaa REST API — واجهة برمجة تطبيقات إحياء
  version: 1.0.0
  description: |
    # Ihyaa (إحياء) — REST API

    منصة إحياء تربط أصحاب الأفكار بالمستثمرين عبر تقييم ذكي بالذكاء الاصطناعي.
    Ihyaa connects idea owners with investors through AI-powered project evaluation.

    ## المصادقة / Authentication
    تسجيل الدخول يمنح توكن Sanctum صالحاً لمدة 24 ساعة يُرسل في الرأس:
    `Authorization: Bearer {token}` — `SRS-NFR-07`.
    Login issues a 24-hour Sanctum token sent via the header above.

    ## الأخطاء / Errors
    جميع الأخطاء تتبع الهيكل الموحد `{code, message, errors}` —
    مع رؤوس `X-RateLimit-*` في كل استجابة و`Retry-After` عند 429 (انظر `rate-limiting-spec.md`).

    ## المصادر المرجعية / References
    - SRS-MVP v3 §5.4.5 (SRS-API-01..49) · `requirements/srs-mvp-v3.md`
    - Rate Limiting: `requirements/rate-limiting-spec.md` (7 مستويات L1..L7)
    - القيم الثابتة: `docs/api/enums.md`

  contact:
    name: Ihyaa Engineering Team
  license:
    name: Proprietary — MVP internal
    identifier: LicenseRef-proprietary
servers:
  - url: https://api.ihyaa.app/api
    description: الإنتاج / Production
  - url: http://127.0.0.1:8000/api
    description: التطوير المحلي / Local development
security:
  - {}
...
paths:
...
  /profile/avatar:
    post:
      tags: [Profile]
      summary: رفع الصورة الشخصية / Upload avatar (SRS-API-11)
      description: |
        رفع صورة شخصية (jpeg/png/webp حتى 2MB) — تُحفظ في Local Disk.
        Rate Limit: upload.file (L5) — 10/دقيقة.
      operationId: uploadAvatar
      x-srs-api: SRS-API-11
      x-rate-limit: "10/دقيقة · user_id · L5"
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar: { type: string, format: binary, description: ملف الصورة / Image file }
      responses:
        "200":
          description: تم الرفع / Uploaded
          content:
            application/json:
              schema:
                type: object
                required: [message, avatar_url]
                properties:
                  message: { type: string }
                  avatar_url: { type: string, format: uri }
              example:
                message: تم تحديث الصورة الشخصية
                avatar_url: https://api.ihyaa.app/storage/avatars/12_1700.jpg
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "422":
          description: صيغة أو حجم غير صالحين / Invalid type or size
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
              examples:
                type: { summary: صيغة مرفوضة, value: { code: FILE_TYPE_NOT_ALLOWED, message: صيغة الملف غير مسموحة } }
                size: { summary: حجم كبير, value: { code: FILE_TOO_LARGE, message: حجم الملف تجاوز الحد المسموح (2MB) } }
        "429":
          $ref: "#/components/responses/RateLimitError"
...
  /projects/{project}/evaluation-status:
    get:
      tags: [Evaluation]
      summary: حالة التقييم / Evaluation status (SRS-API-47)
      description: |
        استعلام حالة التقييم الحالي (processing/completed/failed) — للمالك.
        يُستخدم مع WebSocket لتحديث الواجهة. Rate Limit: shared.read.
      operationId: evaluationStatus
      x-srs-api: SRS-API-47
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ProjectPath"
      responses:
        "200":
          description: الحالة الحالية / Current status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvaluationStatusResponse" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum Token
      description: توكن Sanctum صالح 24 ساعة — يحصل عليه من /login أو /email/verify أو OAuth callback
  parameters:
...
    ProjectPath:
      name: project
      in: path
      required: true
      description: معرّف المشروع / Project ID
      schema: { type: integer, format: int64 }
...
  responses:
    Unauthenticated:
      description: توكن مفقود أو منتهي / Missing or expired token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: UNAUTHENTICATED, message: يجب تسجيل الدخول للوصول إلى هذا المورد }
    Forbidden:
      description: لا صلاحية — دور أو ملكية أو إفصاح / Forbidden — role, ownership or disclosure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: FORBIDDEN, message: ليس لديك صلاحية الوصول إلى هذه الصفحة }
    NotFound:
      description: المورد غير موجود / Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
          example: { code: NOT_FOUND, message: المورد المطلوب غير موجود }
...
    RateLimitError:
      description: تجاوز معدل الطلبات — مع رؤوس X-RateLimit-* وRetry-After / Rate limit exceeded
      headers:
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer } }
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
...
  schemas:
...
    EvaluationStatus:
      type: string
      enum: [processing, completed, failed, partial]
      description: حالة تقييم AI / Evaluation status
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: رمز الخطأ الآلي / Machine-readable code }
        message: { type: string, description: رسالة مترجمة / Localized message }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          description: تفاصيل تحقق لكل حقل / Per-field validation details
...
    RateLimitError:
      type: object
      required: [code, message, retry_after, reset_at]
      properties:
        code: { type: string, const: RATE_LIMIT_EXCEEDED }
        message: { type: string }
        retry_after: { type: integer, description: ثوانٍ حتى المحاولة التالية / Seconds until retry }
        reset_at: { type: integer, description: Unix timestamp لإعادة التعيين / Reset timestamp }
...
    EvaluationStatusResponse:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/EvaluationStatus" }
        evaluation_id: { type: [integer, "null"], format: int64 }
        overall_score: { type: [number, "null"] }
        message: { type: [string, "null"], description: رسالة الحالة (processing/failed) / Status message }
        started_at: { type: [string, "null"], format: date-time }
        eta_seconds: { type: [integer, "null"] }
        last_evaluation_at: { type: [string, "null"], format: date-time }
        next_evaluation_at: { type: [string, "null"], format: date-time }

    # ---------------- Search ----------------
...